+++ devito/passes/clusters/__init__.py
@@ -1,3 +1,4 @@
+from .utils import *  # noqa
 from .aliases import *  # noqa
 from .cse import *  # noqa
 from .factorization import *  # noqa
