+++ devito/passes/clusters/aliases.py
@@ -7,8 +7,7 @@ from devito.ir import (ROUNDABLE, DataSpace, IterationInstance, IterationSpace,
                        Interval, IntervalGroup, LabeledVector, Stencil,
                        detect_accesses, build_intervals)
 from devito.logger import warning
-from devito.passes.clusters.rewriters import dse_pass
-from devito.passes.clusters.utils import make_is_time_invariant
+from devito.passes.clusters.utils import dse_pass, make_is_time_invariant
 from devito.symbolics import estimate_cost, retrieve_indexed
 from devito.types import Array, Eq
