+++ devito/passes/clusters/cse.py
@@ -1,8 +1,7 @@
 from collections import OrderedDict
 
 from devito.ir import DummyEq
-from devito.passes.clusters.rewriters import dse_pass
-from devito.passes.clusters.utils import makeit_ssa
+from devito.passes.clusters.utils import dse_pass, makeit_ssa
 from devito.symbolics import (count, estimate_cost, q_xop, q_leaf, retrieve_terminals,
                               yreplace)
 from devito.tools import DAG, split
