+++ devito/passes/clusters/factorization.py
@@ -1,6 +1,6 @@
 from sympy import Add, Mul, collect, collect_const
 
-from devito.passes.clusters.rewriters import dse_pass
+from devito.passes.clusters.utils import dse_pass
 from devito.symbolics import estimate_cost, retrieve_scalars
 from devito.tools import ReducerMap
