+++ devito/passes/clusters/utils.py
@@ -1,9 +1,32 @@
-from collections import OrderedDict
+from collections import Iterable, OrderedDict
+from functools import wraps
 
 from devito.symbolics import retrieve_terminals
+from devito.tools import flatten, timed_pass
 from devito.types import Dimension, Symbol
 
-__all__ = ['makeit_ssa', 'make_is_time_invariant']
+__all__ = ['dse_pass', 'makeit_ssa', 'make_is_time_invariant']
+
+
+def dse_pass(func):
+    @wraps(func)
+    def wrapper(*args):
+        if timed_pass.is_enabled():
+            maybe_timed = lambda *_args: timed_pass(func, func.__name__)(*_args)
+        else:
+            maybe_timed = lambda *_args: func(*_args)
+        args = list(args)
+        maybe_clusters = args.pop(0)
+        if isinstance(maybe_clusters, Iterable):
+            # Instance method
+            processed = [maybe_timed(c, *args) for c in maybe_clusters]
+        else:
+            # Pure function
+            self = maybe_clusters
+            clusters = args.pop(0)
+            processed = [maybe_timed(self, c, *args) for c in clusters]
+        return flatten(processed)
+    return wrapper
 
 
 def makeit_ssa(exprs):
